Renamed system_input files to input_port_value.
