+++ drake/examples/schunk_wsg/test/schunk_wsg_lcm_test.cc
@@ -6,7 +6,7 @@
 
 #include "drake/lcmt_schunk_wsg_command.hpp"
 #include "drake/systems/analysis/simulator.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 #include "drake/systems/framework/system_output.h"
 
 namespace drake {
