+++ drake/systems/controllers/test/pid_controller_test.cc
@@ -6,7 +6,7 @@
 
 #include "drake/common/eigen_types.h"
 #include "drake/systems/framework/basic_vector.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 using std::make_unique;
