+++ drake/systems/framework/input_port_value.cc
@@ -1,4 +1,4 @@
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 namespace drake {
 namespace systems {
