+++ drake/systems/framework/input_port_value.h
