+++ drake/systems/framework/test/input_port_value_test.cc
@@ -1,4 +1,4 @@
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 #include <memory>
 #include <string>
