+++ drake/systems/plants/spring_mass_system/test/spring_mass_system_test.cc
@@ -10,7 +10,7 @@
 #include "drake/systems/framework/state.h"
 #include "drake/systems/framework/subvector.h"
 #include "drake/systems/framework/system.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 #include "drake/systems/framework/system_output.h"
 #include "drake/systems/framework/vector_base.h"
