+++ drake/systems/primitives/test/constant_vector_source_test.cc
@@ -7,7 +7,7 @@
 #include <gtest/gtest.h>
 
 #include "drake/systems/framework/basic_vector.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 using Eigen::Matrix;
 using std::make_unique;
