+++ drake/systems/primitives/test/gain_test.cc
@@ -6,7 +6,7 @@
 
 #include "drake/common/eigen_types.h"
 #include "drake/systems/framework/basic_vector.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 using Eigen::Vector3d;
 using std::make_unique;
