+++ drake/systems/primitives/test/multiplexer_test.cc
@@ -5,7 +5,7 @@
 #include <gtest/gtest.h>
 
 #include "drake/systems/framework/basic_vector.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 
 using std::make_unique;
