+++ drake/systems/primitives/test/zero_order_hold_test.cc
@@ -9,7 +9,7 @@
 
 #include "drake/common/eigen_types.h"
 #include "drake/systems/framework/basic_vector.h"
-#include "drake/systems/framework/system_input.h"
+#include "drake/systems/framework/input_port_value.h"
 #include "drake/systems/framework/system_output.h"
 
 namespace drake {
